pyproject.toml
[project]
name = "XUnity_Translate_Server"
version = "0.1.0"
description = ""
authors = [{ name = "jp-yendo" }]
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
keywords = ["XUnity", "translate", "server"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "Flask>=3.1.0",
    "openai>=2.0.0",
    "anthropic>=0.70.0",
    "google-genai>=1.49.0",
    "ollama>=0.6.0"
]

[project.urls]
"Homepage" = "https://github.com/jp-yendo/XUnity_Translate_Server"
"Bug Tracker" = "https://github.com/jp-yendo/XUnity_Translate_Server/issues"
"Documentation" = "https://github.com/jp-yendo/XUnity_Translate_Server#readme"

[project.optional-dependencies]
dev = [
    "pylint",
    "pylint-plugin-utils",
    "black"
]
build = [
    "build>=1.0.0",
    "twine>=5.0.0",
    "wheel>=0.42.0"
]

[project.scripts]
XUnity_Translate_Server = "trans_server.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "build>=1.0.0",
    "twine>=5.0.0",
    "wheel>=0.42.0",
]

[tool.black]
line-length = 160
exclude = 'tests/'

[tool.pylint.MASTER]
#load-plugins = ""

[tool.pylint.FORMAT]
max-line-length = 160

[tool.pylint.DESIGN]
max-args = 10
max-locals = 25
max-branches = 20
max-statements = 60

[tool.pylint.messages_control]
disable = [
    "W0718",
]

[tool.setuptools.packages.find]
where = ["src"]
